Renamed scheduler
